requirements.txt
swarms
pydantic
loguru
python-dotenv
swarm-models
numpy
Pillow
openai
qwen-vl-utils
